JPA conflict fixed

+++ src/main/java/com/toolbox/repository/ContactRepository.java
@@ -1,8 +1,9 @@
 package com.toolbox.repository;
+
 import com.toolbox.model.Contact;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ContactRepository extends JpaRepository<Contact,Long>{
+public interface ContactRepository extends JpaRepository<Contact, Integer> {
 }
